feat: retrieve saved code on load
